Add discovery config to the configuration server and configuration server config to Eureka.
Note that we are playing chicken-egg game here, as we cannot register in Eureka if we are not able to get its configuration properties, that are stored in the configuration server.
We are also adding the prefer ip address configuration to avoid problems when running on not real / unproperly configured networks.